refactor: clean up eslint and ts import paths
This commit splits up our big products page route into multiple reusable
components and updates our ESLint and TS import path configurations for
much cleaner and more readable import sections.
