+++ app/entry.client.tsx
@@ -2,7 +2,7 @@ import { StrictMode, startTransition } from 'react';
 import { RemixBrowser } from '@remix-run/react';
 import { hydrateRoot } from 'react-dom/client';
 
-import { logStayCurious } from '~/curious';
+import { logStayCurious } from 'curious';
 
 const hydrate = () => {
   logStayCurious();
