+++ app/routes/healthcheck.tsx
@@ -1,7 +1,7 @@
 // learn more: https://fly.io/docs/reference/configuration/#services-http_checks
 import type { LoaderArgs } from '@remix-run/server-runtime';
 
-import { prisma } from '~/db.server';
+import { prisma } from 'db.server';
 
 export async function loader({ request }: LoaderArgs) {
   const host =
