+++ app/routes/join.tsx
@@ -3,9 +3,10 @@ import type { ActionArgs, LoaderArgs, MetaFunction } from '@remix-run/node';
 import { Form, Link, useActionData, useSearchParams } from '@remix-run/react';
 import { json, redirect } from '@remix-run/node';
 
-import { createUser, getUserByEmail } from '~/models/user.server';
-import { createUserSession, getUserId } from '~/session.server';
-import { safeRedirect, validateEmail, validateUsername } from '~/utils';
+import { createUser, getUserByEmail } from 'models/user.server';
+
+import { createUserSession, getUserId } from 'session.server';
+import { safeRedirect, validateEmail, validateUsername } from 'utils';
 
 export async function loader({ request }: LoaderArgs) {
   const userId = await getUserId(request);
