+++ app/routes/login.tsx
@@ -3,9 +3,10 @@ import type { ActionArgs, LoaderArgs, MetaFunction } from '@remix-run/node';
 import { Form, Link, useActionData, useSearchParams } from '@remix-run/react';
 import { json, redirect } from '@remix-run/node';
 
-import { createUserSession, getUserId } from '~/session.server';
-import { safeRedirect, validateEmail } from '~/utils';
-import { verifyLogin } from '~/models/user.server';
+import { verifyLogin } from 'models/user.server';
+
+import { createUserSession, getUserId } from 'session.server';
+import { safeRedirect, validateEmail } from 'utils';
 
 export async function loader({ request }: LoaderArgs) {
   const userId = await getUserId(request);
