+++ app/routes/logout.tsx
@@ -1,7 +1,7 @@
 import type { ActionArgs } from '@remix-run/node';
 import { redirect } from '@remix-run/node';
 
-import { logout } from '~/session.server';
+import { logout } from 'session.server';
 
 export async function action({ request }: ActionArgs) {
   return logout(request);
