+++ tsconfig.json
@@ -17,8 +17,7 @@
     "strict": true,
     "allowJs": true,
     "forceConsistentCasingInFileNames": true,
-    "baseUrl": ".",
-    "paths": { "~/*": ["./app/*"] },
+    "baseUrl": "app",
     "skipLibCheck": true,
     "noEmit": true // Remix takes care of building everything in `remix build`
   }
